+++ zcash_client_memory/src/wallet_read.rs
@@ -6,14 +6,14 @@ use std::{
 use nonempty::NonEmpty;
 use secrecy::{ExposeSecret, SecretVec};
 use shardtree::store::ShardStore as _;
-#[cfg(feature = "transparent-inputs")]
-use zcash_client_backend::data_api::TransparentBalances;
 use zcash_client_backend::data_api::{
     AddressInfo, BlockMetadata, NullifierQuery, ReceivedTransactionOutput, WalletRead,
     WalletSummary, Zip32Derivation,
     scanning::ScanRange,
     wallet::{ConfirmationsPolicy, TargetHeight},
 };
+#[cfg(feature = "transparent-inputs")]
+use zcash_client_backend::data_api::{TransparentBalances, TransparentKeyOrigin};
 use zcash_client_backend::{
     data_api::{
         Account as _, AccountBalance, AccountSource, Balance, Progress, Ratio, SeedRelevance,
@@ -707,9 +707,12 @@ impl<P: consensus::Parameters> WalletRead for MemoryWalletDb<P> {
         }) {
             if self.utxo_is_spendable(outpoint, target_height, confirmations_policy)? {
                 let address = txo.address;
+                let key_origin = TransparentKeyOrigin::Derived {
+                    scope: txo.key_scope,
+                };
                 let entry = balances
                     .entry(address)
-                    .or_insert((Some(txo.key_scope), Balance::ZERO));
+                    .or_insert((key_origin, Balance::ZERO));
 
                 entry.1.add_spendable_value(txo.txout.value())?;
             }